wip: chat UI refactoring, fix issues
