+++ client/src/components/Chat/Menus/Endpoints/components/EndpointModelItem.tsx
@@ -1,0 +1,95 @@
+import React from 'react';
+import { EModelEndpoint } from 'librechat-data-provider';
+import type { Endpoint } from '../types';
+import { useModelSelectorContext } from '../ModelSelectorContext';
+import { CustomMenuItem as MenuItem } from '../CustomMenu';
+
+interface EndpointModelItemProps {
+  modelId: string | null;
+  endpoint: Endpoint;
+  isSelected: boolean;
+}
+
+export function EndpointModelItem({ modelId, endpoint, isSelected }: EndpointModelItemProps) {
+  const { handleSelectModel } = useModelSelectorContext();
+  let modelName = modelId;
+  const avatarUrl = endpoint?.modelIcons?.[modelId ?? ''] || null;
+
+  // Use custom names if available
+  if (
+    endpoint &&
+    modelId &&
+    endpoint.value === EModelEndpoint.agents &&
+    endpoint.agentNames?.[modelId]
+  ) {
+    modelName = endpoint.agentNames[modelId];
+  } else if (
+    endpoint &&
+    modelId &&
+    endpoint.value === EModelEndpoint.assistants &&
+    endpoint.assistantNames?.[modelId]
+  ) {
+    modelName = endpoint.assistantNames[modelId];
+  }
+
+  return (
+    <MenuItem
+      key={modelId}
+      onClick={() => handleSelectModel(endpoint, modelId ?? '')}
+      className="flex h-8 w-full cursor-pointer items-center justify-start rounded-lg px-3 py-2 text-sm"
+    >
+      <div className="flex items-center gap-2">
+        {avatarUrl ? (
+          <div className="flex h-5 w-5 items-center justify-center overflow-hidden rounded-full">
+            <img src={avatarUrl} alt={modelName ?? ''} className="h-full w-full object-cover" />
+          </div>
+        ) : (endpoint.value === EModelEndpoint.agents ||
+            endpoint.value === EModelEndpoint.assistants) &&
+          endpoint.icon ? (
+            <div className="flex h-5 w-5 items-center justify-center overflow-hidden rounded-full">
+              {endpoint.icon}
+            </div>
+          ) : null}
+        <span>{modelName}</span>
+      </div>
+      {isSelected && (
+        <svg
+          width="16"
+          height="16"
+          viewBox="0 0 24 24"
+          fill="none"
+          xmlns="http://www.w3.org/2000/svg"
+          className="ml-auto block"
+        >
+          <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M2 12C2 6.47715 6.47715 2 12 2C17.5228 2 22 6.47715 22 12C22 17.5228 17.5228 22 12 22C6.47715 22 2 17.5228 2 12ZM16.0755 7.93219C16.5272 8.25003 16.6356 8.87383 16.3178 9.32549L11.5678 16.0755C11.3931 16.3237 11.1152 16.4792 10.8123 16.4981C10.5093 16.517 10.2142 16.3973 10.0101 16.1727L7.51006 13.4227C7.13855 13.014 7.16867 12.3816 7.57733 12.0101C7.98598 11.6386 8.61843 11.6687 8.98994 12.0773L10.6504 13.9039L14.6822 8.17451C15 7.72284 15.6238 7.61436 16.0755 7.93219Z"
+            fill="currentColor"
+          />
+        </svg>
+      )}
+    </MenuItem>
+  );
+}
+
+export function renderEndpointModels(
+  endpoint: Endpoint | null,
+  models: string[],
+  selectedModel: string | null,
+  filteredModels?: string[],
+) {
+  const modelsToRender = filteredModels || models;
+
+  return modelsToRender.map(
+    (modelId) =>
+      endpoint && (
+        <EndpointModelItem
+          key={modelId}
+          modelId={modelId}
+          endpoint={endpoint}
+          isSelected={selectedModel === modelId}
+        />
+      ),
+  );
+}
